+++ src/mat2.cpp
@@ -3,7 +3,7 @@
 namespace tdm
 {
 
-Mat2 Mat2::Zero = Mat2{0.0f, 0.0f, 0.0f, 0.0f};
-Mat2 Mat2::Identity = Mat2{1.0f, 0.0f, 0.0f, 1.0f};
+const Mat2 Mat2::Zero = Mat2{0.0f, 0.0f, 0.0f, 0.0f};
+const Mat2 Mat2::Identity = Mat2{1.0f, 0.0f, 0.0f, 1.0f};
 
 } // namespace tdm