+++ src/vec2.cpp
@@ -2,11 +2,11 @@
 namespace tdm
 {
 
-Vec2 Vec2::Zero = Vec2(0.0f, 0.0f);
+const Vec2 Vec2::Zero = Vec2(0.0f, 0.0f);
 
-Vec2 Vec2::Right = Vec2(1.0f, 0.0f);
-Vec2 Vec2::Left = Vec2(-1.0f, 0.0f);
-Vec2 Vec2::Up = Vec2(0.0f, 1.0f);
-Vec2 Vec2::Down = Vec2(0.0f, -1.0f);
+const Vec2 Vec2::Right = Vec2(1.0f, 0.0f);
+const Vec2 Vec2::Left = Vec2(-1.0f, 0.0f);
+const Vec2 Vec2::Up = Vec2(0.0f, 1.0f);
+const Vec2 Vec2::Down = Vec2(0.0f, -1.0f);
 
 } // namespace tdm